リファクタリング #47

+++ KakeiboApp/Model/CalendarModel.swift
@@ -1,5 +1,5 @@
 //
-//  ClalendarModel.swift
+//  CalendarModel.swift
 //  KakeiboApp
 //
 //  Created by Hyakusetufutou on 2025/11/06
@@ -11,7 +11,7 @@ import Foundation
 struct DailySummary {
     let id = UUID()
     let date: Date
-    let income: Double
-    let expense: Double
+    let income: Decimal
+    let expense: Decimal
     let transactions: [TransactionModel]
 }
